SCAL-1051: Fast visualization Added visualization for FAST method && refactored modal and plugin in order to work with many sensitivity analysis methods from ES
