--- app/visualisation_methods/morris/chart.html.haml
@@ -1,14 +1,0 @@
-%section.panel.radius.plot
-  %h3.subheader Morris design - #{params["output"]} - (chart id: #{params["chart_id"].to_i + 1})
-  .row
-  %span.button.hide{id: "morris_hide_#{params['chart_id']}"} Hide chart
-  %br
-    .chart
-
-:javascript
-  $(function() {
-    var div_id = "#morris_chart_#{params['chart_id']}";
-    $("#morris_hide_#{params['chart_id']}").on('click', function() {
-      $(div_id).remove();
-    });
-  });
