+++ app/visualisation_methods/sensitivity/info.json
@@ -1,0 +1,7 @@
+{
+  "name": "Sensitivity design",
+  "id": "sensitivity",
+  "group": "params",
+  "image": "sensitivity_desing_icon.png",
+  "description":  "Shows results of sensitivity analysis with conducted specific method (automatic experiment type)."
+}
