+++ config/methods.json
@@ -1,5 +1,5 @@
 {
-    "methods": ["lindev", "three_d", "pareto", "interaction", "dendrogram", "kmeans", "morris"],
+    "methods": ["lindev", "three_d", "pareto", "interaction", "dendrogram", "kmeans", "sensitivity"],
 
     "groups": {
